+++ src/lna_hrbf_rcpp.cpp
@@ -7,10 +7,10 @@ using namespace Rcpp;
 // [[Rcpp::depends(RcppEigen)]]
 
 // [[Rcpp::export]]
-Eigen::SparseMatrix<float> hrbf_atoms_rcpp(
-    const Eigen::Map<Eigen::MatrixXf> mask_xyz_world,
-    const Eigen::Map<Eigen::MatrixXf> centres_xyz_world,
-    const Eigen::Map<Eigen::VectorXf> sigma_vec_mm,
+Eigen::SparseMatrix<double> hrbf_atoms_rcpp(
+    const Eigen::Map<Eigen::MatrixXd> mask_xyz_world,
+    const Eigen::Map<Eigen::MatrixXd> centres_xyz_world,
+    const Eigen::Map<Eigen::VectorXd> sigma_vec_mm,
     std::string kernel_type,
     double value_threshold = 1e-8)
 {
@@ -21,33 +21,33 @@ Eigen::SparseMatrix<float> hrbf_atoms_rcpp(
     }
     const bool gaussian = (kernel_type == "gaussian");
 
-    std::vector< Eigen::Triplet<float> > triplets;
+    std::vector< Eigen::Triplet<double> > triplets;
 #ifdef _OPENMP
     int nthreads = omp_get_max_threads();
-    std::vector< std::vector<Eigen::Triplet<float>> > thread_triplets(nthreads);
+    std::vector< std::vector<Eigen::Triplet<double>> > thread_triplets(nthreads);
 #pragma omp parallel
     {
         int tid = omp_get_thread_num();
-        std::vector<Eigen::Triplet<float>>& local = thread_triplets[tid];
+        std::vector<Eigen::Triplet<double>>& local = thread_triplets[tid];
 #pragma omp for schedule(static)
         for (int k = 0; k < K; ++k) {
-            float sigma = sigma_vec_mm[k];
-            float inv_two_sigma2 = 1.0f / (2.0f * sigma * sigma);
-            float inv_sigma = 1.0f / sigma;
+            double sigma = sigma_vec_mm[k];
+            double inv_two_sigma2 = 1.0 / (2.0 * sigma * sigma);
+            double inv_sigma = 1.0 / sigma;
             for (int n = 0; n < N; ++n) {
-                float dx = mask_xyz_world(n,0) - centres_xyz_world(k,0);
-                float dy = mask_xyz_world(n,1) - centres_xyz_world(k,1);
-                float dz = mask_xyz_world(n,2) - centres_xyz_world(k,2);
-                float dist2 = dx*dx + dy*dy + dz*dz;
-                float value;
+                double dx = mask_xyz_world(n,0) - centres_xyz_world(k,0);
+                double dy = mask_xyz_world(n,1) - centres_xyz_world(k,1);
+                double dz = mask_xyz_world(n,2) - centres_xyz_world(k,2);
+                double dist2 = dx*dx + dy*dy + dz*dz;
+                double value;
                 if (gaussian) {
                     value = std::exp(-dist2 * inv_two_sigma2);
                 } else {
-                    float dist = std::sqrt(dist2) * inv_sigma;
-                    if (dist >= 1.0f) continue;
-                    float base = 1.0f - dist;
-                    value = std::pow(base,8.0f) * (32.0f*dist*dist*dist +
-                                                  25.0f*dist*dist + 8.0f*dist + 1.0f);
+                    double dist = std::sqrt(dist2) * inv_sigma;
+                    if (dist >= 1.0) continue;
+                    double base = 1.0 - dist;
+                    value = std::pow(base,8.0f) * (32.0*dist*dist*dist +
+                                                  25.0f*dist*dist + 8.0f*dist + 1.0);
                 }
                 if (std::fabs(value) > value_threshold) {
                     local.emplace_back(k, n, value);
@@ -61,23 +61,23 @@ Eigen::SparseMatrix<float> hrbf_atoms_rcpp(
 #else
     triplets.reserve(static_cast<size_t>(N) * 4);
     for (int k = 0; k < K; ++k) {
-        float sigma = sigma_vec_mm[k];
-        float inv_two_sigma2 = 1.0f / (2.0f * sigma * sigma);
-        float inv_sigma = 1.0f / sigma;
+        double sigma = sigma_vec_mm[k];
+        double inv_two_sigma2 = 1.0 / (2.0 * sigma * sigma);
+        double inv_sigma = 1.0 / sigma;
         for (int n = 0; n < N; ++n) {
-            float dx = mask_xyz_world(n,0) - centres_xyz_world(k,0);
-            float dy = mask_xyz_world(n,1) - centres_xyz_world(k,1);
-            float dz = mask_xyz_world(n,2) - centres_xyz_world(k,2);
-            float dist2 = dx*dx + dy*dy + dz*dz;
-            float value;
+            double dx = mask_xyz_world(n,0) - centres_xyz_world(k,0);
+            double dy = mask_xyz_world(n,1) - centres_xyz_world(k,1);
+            double dz = mask_xyz_world(n,2) - centres_xyz_world(k,2);
+            double dist2 = dx*dx + dy*dy + dz*dz;
+            double value;
             if (gaussian) {
                 value = std::exp(-dist2 * inv_two_sigma2);
             } else {
-                float dist = std::sqrt(dist2) * inv_sigma;
-                if (dist >= 1.0f) continue;
-                float base = 1.0f - dist;
-                value = std::pow(base,8.0f) * (32.0f*dist*dist*dist +
-                                              25.0f*dist*dist + 8.0f*dist + 1.0f);
+                double dist = std::sqrt(dist2) * inv_sigma;
+                if (dist >= 1.0) continue;
+                double base = 1.0 - dist;
+                value = std::pow(base,8.0f) * (32.0*dist*dist*dist +
+                                              25.0f*dist*dist + 8.0f*dist + 1.0);
             }
             if (std::fabs(value) > value_threshold) {
                 triplets.emplace_back(k, n, value);
@@ -85,7 +85,7 @@ Eigen::SparseMatrix<float> hrbf_atoms_rcpp(
         }
     }
 #endif
-    Eigen::SparseMatrix<float> result(K, N);
+    Eigen::SparseMatrix<double> result(K, N);
     if (!triplets.empty()) {
         result.setFromTriplets(triplets.begin(), triplets.end());
     }
